+++ src/api/index.ts
@@ -15,6 +15,7 @@ import {
   SerieAPlayer,
   Club,
 } from "../types";
+import { CoppaRounds, Round } from "../types/coppa";
 
 const { language } = getPreferenceValues();
 const cache = new Cache();
@@ -99,13 +100,20 @@ export const getStandings = async (season: string): Promise<Standing[]> => {
 
 export const getMatches = async (
   season: string,
-  matchday: number
+  params: any
 ): Promise<Match[]> => {
   const [title, season_id] = season.split("_");
 
   const config: AxiosRequestConfig = {
     method: "GET",
-    url: `${endpoint}/stats/live/match?extra_link=&order=oldest&lang=en&season_id=${season_id}&match_day_id=${matchday}`,
+    url: `${endpoint}/stats/live/match`,
+    params: {
+      extra_link: "",
+      order: "oldest",
+      lang: "en",
+      season_id,
+      ...params,
+    },
   };
 
   try {
@@ -183,5 +191,24 @@ export const getClub = async (slug: string): Promise<Club | undefined> => {
     showFailureToast();
 
     return undefined;
+  }
+};
+
+export const getCoppaRounds = async (season: string): Promise<Round[]> => {
+  const [title, seasonId] = season.split("_");
+
+  const config: AxiosRequestConfig = {
+    method: "GET",
+    url: `https://www.legaseriea.it/api/season/${seasonId}/championship/CPITA/rounds?lang=en`,
+  };
+
+  try {
+    const { data }: AxiosResponse<CoppaRounds> = await axios(config);
+
+    return data.data;
+  } catch (e) {
+    showFailureToast();
+
+    return [];
   }
 };