+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Improvements] - 2023
 - Added Club details
 - Refine Fixtures & Results (live fixtures status, broadcasters)
+- Added Coppa Italia
 
 ## [Initial Version] - 2023-02-10